concurrent prediction support

+++ events/messageCreate.ts
@@ -1,15 +1,16 @@
 import { Message } from "discord.js";
-import readConfig from "../utils/readConfig";
+import getDbOptions from "../utils/getDbOptions";
+import { createConnection } from "mysql2/promise";
 
 const messageCreate = async (message: Message) => {
-  const config = readConfig();
-  if (config.predictionChannelId) {
-    if (
-      message.channelId === config.predictionChannelId &&
-      !message.member?.permissions.has("ManageMessages")
-    )
-      message.delete();
-  }
+  const db = await createConnection(getDbOptions());
+  const [rows] = await db.execute(
+    "SELECT * FROM predictions WHERE channelId = ? LIMIT 1",
+    [message.channelId]
+  );
+  db.end();
+  if (!Array.isArray(rows)) return;
+  if (!message.member?.permissions.has("ManageMessages")) message.delete();
 };
 
 module.exports = messageCreate;
